adding a way to mock the v2.5 platform API.
The har parser creates the path and json files that allows the server
to respond to /api/imls/* requests from a captured .har file.

I used the OERX on our existing site, saved the .har file, used this
utility to create static json files and created a
middleware (server.helpers.MocAPI) that maps the API endpoints to the
saved json files.

IMPORTANT: you don't need to change what's added in this commit to
start to override the API endpoints you want to implement for real. As
long as the route and the handler appears in server.main before the
MocAPI handler, it will get to process the incoming request and the
MocAPI handler is only a fallback for API endpoints that haven't been
implemented yet.